refactor: move `QemuNetworkInterfaceID` to seperate file
