+++ proxmox/config_qemu_test.go
@@ -7646,22 +7646,3 @@ func Test_ConfigQemu_Validate(t *testing.T) {
 		}
 	}
 }
-
-func Test_QemuNetworkInterfaceID_Validate(t *testing.T) {
-	tests := []struct {
-		name   string
-		input  QemuNetworkInterfaceID
-		output error
-	}{
-		{name: "Valid",
-			input: QemuNetworkInterfaceID0},
-		{name: "Invalid",
-			input:  32,
-			output: errors.New(QemuNetworkInterfaceID_Error_Invalid)},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			require.Equal(t, test.output, test.input.Validate())
-		})
-	}
-}
